Refactor: Correct method calls and type conversions
This commit includes several refactoring changes:
- Corrected method names in `CaseViewModel` when interacting with `caseRepository`.
  - `importSpreadsheetAndStore` changed to `importSpreadsheet`.
  - `createNewCaseWithDetails` changed to `createCase`.
  - `archiveExistingCase` changed to `archiveCase`.
- In `AddonsBrowserScreen`, converted `Double` rating values to `Float` for `AddonCard`.
- Added a new string resource `script_editor_placeholder` in `strings.xml`.
- In `TranscriptionService`, corrected the access token retrieval from `credential.getCredential().getAccessToken()` to `credential.credential.accessToken`.
- Updated `EvidenceDetailsViewModelFactory` to accept `Application` as a constructor parameter and pass it to `EvidenceDetailsViewModel`.

+++ app/src/main/java/com/hereliesaz/lexorcist/ui/AddonsBrowserScreen.kt
@@ -50,7 +50,7 @@ fun AddonsBrowserScreen(
                         name = script.name,
                         description = script.description,
                         author = script.author,
-                        rating = script.rating,
+                        rating = script.rating.toFloat(), // Converted Double to Float
                         onRate = { rating ->
                             viewModel.rateAddon(script.id, rating, "Script")
                         }
@@ -64,7 +64,7 @@ fun AddonsBrowserScreen(
                         name = template.name,
                         description = template.description,
                         author = template.author,
-                        rating = template.rating,
+                        rating = template.rating.toFloat(), // Converted Double to Float
                         onRate = { rating ->
                             viewModel.rateAddon(template.id, rating, "Template")
                         }
